Split state from analysis
