+++ analysis/src/abstract_interpretation/pointwise_state.rs
@@ -4,7 +4,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use crate::AbstractState;
+use crate::abstract_interpretation::AbstractState;
 use rustc_data_structures::fx::FxHashMap;
 use rustc_middle::mir;
 use serde::{ser::SerializeMap, Serialize, Serializer};
