+++ analysis/src/analysis_error.rs
@@ -4,7 +4,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-use crate::serialization_utils::location_to_stmt_str;
+use crate::mir_utils::location_to_stmt_str;
 use rustc_middle::mir;
 
 #[derive(Debug)]
