+++ analysis/src/mir_utils.rs
@@ -14,6 +14,19 @@ use rustc_middle::{
     ty::{self, TyCtxt},
 };
 
+/// Convert a `location` to a string representing the statement or terminator at that `location`
+pub fn location_to_stmt_str(location: mir::Location, mir: &mir::Body) -> String {
+    let bb_mir = &mir[location.block];
+    if location.statement_index < bb_mir.statements.len() {
+        let stmt = &bb_mir.statements[location.statement_index];
+        format!("{:?}", stmt)
+    } else {
+        // location = terminator
+        let terminator = bb_mir.terminator();
+        format!("{:?}", terminator.kind)
+    }
+}
+
 /// Check if the place `potential_prefix` is a prefix of `place`. For example:
 ///
 /// +   `is_prefix(x.f, x.f) == true`
